requirements.txt
alembic==1.15.2
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
click==8.1.8
cloudinary==1.44.1
colorama==0.4.6
cssselect2==0.8.0
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyMuPDF==1.25.3
pyphen==0.17.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.2.1
pytz==2025.1
reportlab==4.3.1
six==1.17.0
SQLAlchemy==2.0.40
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
XlsxWriter==3.2.3
zopfli==0.2.3.post1
